attack system part 1 done

+++ Assets/Scripts/Weapon.cs
@@ -7,7 +7,7 @@ using UnityEditor.Animations;
 public class WeaponScript : MonoBehaviour
 {
     public int damage;
-    public int weaponRange;
+    public float weaponRange;
     public AnimatorController animator;
     public bool attachLeft;
 }
